requirements.txt
# data processing
pandas==1.5.3

# Asana py client
asana==3.0.0

# load env
python-dotenv==0.21.1

# google apis
google-api-core==2.3.2
google-api-python-client==2.33.0
google-cloud-bigquery==3.1.0
pandas-gbq==0.19.1